Crate monarch_types

Re-exports

pub use nccl_types::NcclUniqueId;

pub use nccl_types::ReduceOp;

pub use nccl_types::UniqueId;

Modules

nccl_types

GPU-independent types for NCCL-based communication.

Macros

py_global

Macro to generate a Python object lookup function with caching

py_module_add_function

Macro to register a function to a Python module.

Structs

PickledPyObject

PyTree

A Rust wrapper around A PyTorch pytree, which can be serialized and sent across the wire. https://github.com/pytorch/pytorch/blob/main/torch/utils/_pytree.py

SerializablePyErr

A wrapper around PyErr that contains a serialized traceback.

Traits

MapPyErr

TryIntoPyObjectUnsafe

A variant of pyo3::IntoPyObject used to wrap unsafe impls and propagates the unsafety to the caller.