monarch_types

Struct PyTree

pub struct PyTree<T> { /* private fields */ }

A Rust wrapper around A PyTorch pytree, which can be serialized and sent across the wire. https://github.com/pytorch/pytorch/blob/main/torch/utils/_pytree.py

Implementations

impl<T> PyTree<T>

pub fn is_leaf(&self) -> bool

pub fn into_leaf(self) -> Option<T>

pub fn leaves(&self) -> &[T]

pub fn into_leaves(self) -> Vec<T>

pub fn for_each<F>(&self, func: F)
where F: FnMut(&T),

pub fn iter<'a>(&'a self) -> Iter<'a, T>

pub fn iter_mut<'a>(&'a mut self) -> IterMut<'a, T>

pub fn into_map<U, F>(self, func: F) -> PyTree<U>
where F: FnMut(T) -> U,

Map leaf values with the given func.

pub fn try_map<U, F, E>(&self, func: F) -> Result<PyTree<U>, E>
where F: FnMut(&T) -> Result<U, E>,

Map the leaves of the pytree with the given fallible callback.

pub fn try_into_map<U, F, E>(self, func: F) -> Result<PyTree<U>, E>
where F: FnMut(T) -> Result<U, E>,

Map the leaves of the pytree with the given fallible callback.

impl<'a, T: FromPyObject<'a>> PyTree<T>

pub fn flatten(tree: &Bound<'a, PyAny>) -> PyResult<Self>